% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.R
\name{aspect}
\alias{aspect}
\title{Calculate topographic aspect from a DEM}
\usage{
aspect(x, na_flag = -9999)
}
\arguments{
\item{x}{A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster, representing a DEM.}

\item{na_flag}{Value use to mark \code{NA} values in C code. This should be set to a value which is guaranteed to be absent from the raster (default is \code{-9999}).}
}
\value{
A \code{stars} raster with topographic slope, i.e., the azimuth where the terrain is tilted towards, in decimal degrees (0-360) clockwise from north. Flat terrain, i.e., when all values in the neighborhood are equal, gets the value of \code{-1}.
}
\description{
Calculates topographic aspect given a Digital Elevation Model (DEM) raster
}
\note{
Aspect calculation results in \code{NA} when at least one of the cell neighbors is \code{NA}, including the outermost rows and columns.
}
\examples{
# Small example
data(dem)
aspect = aspect(dem)
r = c(dem, round(aspect, 1), along = 3)
r = st_set_dimensions(r, 3, values = c("input", "aspect"))
plot(r, text_values = TRUE, breaks = "equal", col = hcl.colors(11, "Spectral"))

\dontrun{

# Larger example
data(carmel)
carmel1 = aspect(carmel)
r = c(carmel, round(carmel1, 1), along = 3)
r = st_set_dimensions(r, 3, values = c("input", "aspect"))
plot(r, breaks = "equal", hcl.colors(11, "Spectral"))

}

}
\references{
The topographic aspect algorithm is based on the \emph{How aspect works} article in the ArcGIS documentation:

\url{https://desktop.arcgis.com/en/arcmap/10.3/tools/spatial-analyst-toolbox/how-aspect-works.htm}
}
