% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carmel}
\alias{carmel}
\title{Digital Elevation Model of Mount Carmel}
\format{
A \code{stars} object with 1 attribute:
\describe{
  \item{elevation}{Elevation above sea level, in meters}
}
}
\usage{
carmel
}
\description{
A \code{stars} object representing a Digital Elevation Model (DEM) Digital Elevation Model of Mount Carmel, at 90m resolution
}
\examples{
plot(carmel, breaks = "equal", col = terrain.colors(11))
}
\keyword{datasets}
