% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\title{Calculate the Convergence Index (CI) from a slope raster}
\usage{
CI(x, k, na.rm = FALSE, na_flag = -9999)
}
\arguments{
\item{x}{A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster, representing slope in decimal degrees (possibly including \code{-1} to specify flat terrain).}

\item{k}{k Neighborhood size around focal cell. Must be an odd number. For example, \code{k=3} implies a 3*3 neighborhood.}

\item{na_flag}{Value use to mark \code{NA} values in C code. This should be set to a value which is guaranteed to be absent from the raster (default is \code{-9999}).}
}
\value{
The filtered \code{stars} raster
}
\description{
Calculates the Convergence Index (CI) given a topographic slope raster
}
\note{
The raster is "padded" with \code{(k-1)/2} more rows and columns of \code{NA} values on all sides, so that the neigborhood of the outermost rows and columns is still a complete neighborhood. Those rows and columns are removed from the final result before returning it.
Aspect values of \code{-1}, specifying flat terrain, are assigned with a CI value of \code{0} regardless of their neighbor values.
}
\examples{
# Small example
data(dem)
dem_asp = aspect(dem)
dem_ci = CI(dem_asp, k = 3)
r = c(dem, round(dem_ci, 1), along = 3)
r = st_set_dimensions(r, 3, values = c("input", "CI (k=3)"))
plot(r, text_values = TRUE, breaks = "equal", col = terrain.colors(10))

\dontrun{

# Larger example
data(golan)
golan_asp = aspect(golan)
golan_ci = CI(golan_asp, k = 25)
plot(golan_ci)

}

}
\references{
The Convergence Index algorithm is described in:

Thommeret, N., Bailly, J. S., & Puech, C. (2010). Extraction of thalweg networks from DTMs: application to badlands.
}
