% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsat}
\alias{landsat}
\title{RGB image of Mount Carmel}
\format{
A \code{stars} object with 1 attribute:
\describe{
  \item{refl}{Reflectance, numeric value between 0 and 1}
}
}
\usage{
landsat
}
\description{
A \code{stars} object representing an RGB image of part of Mount Carmel, at 30m resolution. The data source is Landsat-8 Surface Reflectance product.
}
\keyword{datasets}
