% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_values_matrix.R
\name{layer_values_matrix}
\alias{layer_values_matrix}
\title{Get \code{stars} layer values as matrix}
\usage{
layer_values_matrix(r)
}
\arguments{
\item{r}{A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster}
}
\value{
A \code{matrix} with the layer values, having the same orientation as the raster (i.e., rows represent the y-axis and columns represent the x-axis)
}
\description{
Extracts the values of a single layer in a \code{stars} object to a \code{matrix}.
}
\examples{
data(dem)
m = layer_values_matrix(dem)
m

}
