% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_values_vector.R
\name{layer_values_vector}
\alias{layer_values_vector}
\title{Get \code{stars} layer values as vector}
\usage{
layer_values_vector(r)
}
\arguments{
\item{r}{A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster}
}
\value{
A vector with the layer values, ordered from top-left corner to the right.
}
\description{
Extracts the values of a single layer in a \code{stars} object to a vector. Cell values are ordered from top-left corner to the right.
}
\examples{
data(dem)
v = layer_values_vector(dem)
v

}
