% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_extend.R
\name{matrix_extend}
\alias{matrix_extend}
\title{Extend matrix}
\usage{
matrix_extend(m, n = 1, fill = NA)
}
\arguments{
\item{m}{A \code{matrix}}

\item{n}{By how many rows/columns to extend the matrix on each side?}

\item{fill}{Fill value (default is \code{NA})}
}
\value{
An extended \code{matrix}
}
\description{
Adds \code{n} rows and columns with \code{NA} values on all sides of a \code{matrix}.
}
\examples{
m = matrix(1:6, nrow = 2, ncol = 3)
m
matrix_extend(m, 1)
matrix_extend(m, 2)
matrix_extend(m, 3)
}
