% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_trim.R
\name{matrix_trim}
\alias{matrix_trim}
\title{Trim matrix}
\usage{
matrix_trim(m, n = 1)
}
\arguments{
\item{m}{A \code{matrix}}

\item{n}{By how many rows/columns to trim the matrix on each side?}
}
\value{
A trimmed \code{matrix}, or \code{NULL} if trimming results in an "empty" matrix.
}
\description{
Removes \code{n} rows and columns with \code{NA} values on all sides of a \code{matrix}.
}
\examples{
m = matrix(1:80, nrow = 8, ncol = 10)
m
matrix_trim(m, 1)
matrix_trim(m, 2)
matrix_trim(m, 3)
matrix_trim(m, 4)
}
