% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_value}
\alias{mode_value}
\title{Mode}
\usage{
mode_value(x, na_flag = -9999)
}
\arguments{
\item{x}{A \code{numeric} or \code{logical} vector}

\item{na_flag}{Value used to mark \code{NA} values in C code. This should be set to a value which is guaranteed to be absent from the input vector \code{x} (default is \code{-9999}).}
}
\value{
The mode, \code{numeric} vector of length 1
}
\description{
Find the mode (i.e., most common value) in a numeric vector. In case of ties, the first encountered value is returned.
}
\examples{
x = c(3, 2, 5, 5, 3, 10, 2, 5)
mode_value(x)

}
