% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_normalize_2d.R
\name{normalize_2d}
\alias{normalize_2d}
\title{Normalize a 2D 'stars' object}
\usage{
normalize_2d(x)
}
\arguments{
\item{x}{A \code{stars} object}
}
\value{
A new \code{stars} object
}
\description{
Check, and possibly correct, that the input \code{stars} object:
\itemize{
\item{Has exactly one attribute.}
\item{Has exactly two dimensions.}
\item{The dimensions are spatial, named \code{x} and {y} (in that order).}
}
}
\examples{
# Small example
data(dem)
normalize_2d(dem)
}
