% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_normalize_3d.R
\name{normalize_3d}
\alias{normalize_3d}
\title{Normalize a 3D 'stars' object}
\usage{
normalize_3d(x)
}
\arguments{
\item{x}{A \code{stars} object}
}
\value{
A new \code{stars} object
}
\description{
Check, and possibly correct, that the input \code{stars} object:
\itemize{
\item{Has exactly one attribute.}
\item{Has exactly three dimensions.}
\item{The first two dimensions are spatial, named \code{x} and \code{y} (in that order).}
}
}
\examples{
# Small example
data(landsat)
normalize_3d(landsat)
}
