% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsExtra-package.R
\docType{package}
\name{starsExtra-package}
\alias{starsExtra}
\alias{starsExtra-package}
\title{starsExtra: Miscellaneous Functions for Working with 'stars' Rasters}
\description{
Miscellaneous functions for working with 'stars' objects, mainly single-band rasters. Currently includes functions for: (1) focal filtering, (2) detrending of Digital Elevation Models, (3) calculating flow length, (4) calculating the Convergence Index, (5) calculating topographic aspect and topographic slope.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://michaeldorman.github.io/starsExtra/}
  \item \url{https://github.com/michaeldorman/starsExtra/}
  \item Report bugs at \url{https://github.com/michaeldorman/starsExtra/issues/}
}

}
\author{
\strong{Maintainer}: Michael Dorman \email{dorman@post.bgu.ac.il}

Authors:
\itemize{
  \item Michael Dorman \email{dorman@post.bgu.ac.il}
}

}
\keyword{internal}
