% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Remove empty outer rows and columns}
\usage{
trim(x)
}
\arguments{
\item{x}{A two-dimensional \code{stars} object}
}
\value{
A new \code{stars} object with empty outer rows and columns removed
}
\description{
Removes complete outer rows and columns which have \code{NA} values.
}
\examples{
# Small example
data(dem)
dem[[1]][1,] = NA
dem1 = trim(dem)
}
