% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim2.R
\name{trim2}
\alias{trim2}
\title{Remove empty outer rows and columns}
\usage{
trim2(x)
}
\arguments{
\item{x}{A two-dimensional \code{stars} object}
}
\value{
A new \code{stars} object with empty outer rows and columns removed
}
\description{
Removes complete outer rows and columns which have \code{NA} values.
}
\examples{
# Single-band example
data(dem)
dem[[1]][1,] = NA
dem1 = trim2(dem)

# Multi-band example
data(landsat)
landsat[[1]][1:100,,] = NA
landsat1 = trim2(landsat)

}
