% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsExtra-package.R
\docType{package}
\name{starsExtra-package}
\alias{starsExtra}
\alias{starsExtra-package}
\title{starsExtra: Miscellaneous functions for working with 'stars' rasters}
\description{
Miscellaneous functions for working with 'stars' objects, mainly single-band rasters. Currently includes functions for: (1) extracting raster values as vector or as matrix, (2) applying focal filter.
}
\author{
\strong{Maintainer}: Michael Dorman \email{dorman@post.bgu.ac.il}

}
\keyword{internal}
