% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsExtra-package.R
\docType{package}
\name{starsExtra-package}
\alias{starsExtra}
\alias{starsExtra-package}
\title{starsExtra: Miscellaneous Functions for Working with 'stars' Rasters}
\description{
Miscellaneous functions for working with 'stars' objects, mainly single-band rasters. Currently includes functions for: (1) focal filtering, (2) detrending of Digital Elevation Models, and (3) calculating flow length.
}
\author{
\strong{Maintainer}: Michael Dorman \email{dorman@post.bgu.ac.il}

}
\keyword{internal}
